requirements.txt
tweepy==4.10.1
aiohttp==3.8.1
python-dotenv
schedule